requirements.txt
openai~=0.16.0
backoff~=1.11.1
python-dotenv~=0.20.0
ratelimit~=2.2.1
tomli~=2.0.1
docker~=5.0.3
typer~=0.4.0
tqdm~=4.63.1
list2csv~=1.3.6
